Signal: make it work with native JSON-RPC endpoints of the signal cli package.
The currently used protocols are partly outdated and refer to

https://gitlab.com/morph027/signal-cli-dbus-rest-api
https://gitlab.com/morph027/python-signal-cli-rest-api

which are considered obsolete by its author in favour of the the
JSON-RPC interface of the signal-cli package.

There is also another Docker project

https://github.com/bbernhard/signal-cli-rest-api

which seemingly uses the API used by the obsolete "morph027" packages.
However, as signal-cli already supports a remote API: why not use it,
instead of having to install yet another package?

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>